requirements.txt
streamlit
deepface
tf-keras
tensorflow
opencv-python-headless
pillow
numpy